NFTCreation_v0 is initial version of NFT minting
collection.sol is utilizing ERC721 standard with the function of singlemint(), batchmint(), initialize(). The view function and event is still needed to be added afterward.
collection_factory.sol is utilizing the create2 feature to make instances of NFT.
the above contracts will be tested afterward to make NFT minting from users call is working.